Combine memory tracker into separate header file
And add correct support for aligned allocations
